feat : DeleteProduct 구현 및 쓸모없는 코드 삭제
