+++ OpenMarket/OpenMarket/Model/UserInformation.swift
@@ -5,8 +5,11 @@
 //  Created by dudu, safari on 2022/05/27.
 //
 
+import Foundation
+
 enum UserInformation {
     static let id = 79
     static let identifier = "680a5dc5-d1b8-11ec-9676-819f7cbd69a2"
     static let password = "dxx7xmmtfz"
+    static let boundary = UUID().uuidString
 }
