+++ OpenMarket/OpenMarket/Network/HTTPMethod.swift
@@ -1,0 +1,15 @@
+//
+//  HTTPMethod.swift
+//  OpenMarket
+//
+//  Created by dudu, safari on 2022/05/16.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case patch = "PATCH"
+    case delete = "DELETE"
+}
