+++ OpenMarket/OpenMarket/Network/NetworkManager.swift
@@ -45,35 +45,6 @@ struct NetworkManager {
         }.resume()
     }
     
-    func request<T: Decodable>(endPoint: EndPoint, completion: @escaping (Result<T, NetworkError>) -> Void) {
-        guard let urlRequst = endPoint.urlRequest else {
-            completion(.failure(.urlError))
-            return
-        }
-        
-        session.dataTask(with: urlRequst) { data, response, error in
-            guard let statusCode = (response as? HTTPURLResponse)?.statusCode,
-                  (200..<300).contains(statusCode),
-                  error == nil else {
-                completion(.failure(.severError))
-                return
-            }
-            
-            guard let data = data else {
-                completion(.failure(.dataError))
-                return
-            }
-            
-            do {
-                let result = try JSONDecoder().decode(T.self, from: data)
-                completion(.success(result))
-            } catch {
-                completion(.failure(.jsonError))
-                return
-            }
-        }.resume()
-    }
-    
     func request<T: Decodable>(api: APIable, completion: @escaping (Result<T, NetworkError>) -> Void) {
         guard let urlRequest = api.makeURLRequest() else {
             completion(.failure(.urlError))
